task/types: split TaskItem to types file
Resolves a circular dependency.
