+++ frontend/src/app/tasks/tasks.service.ts
@@ -5,22 +5,8 @@ import {
   ServerBucket,
   ServerResponse
 } from './task-server.service';
+import { TaskItem, TaskPriorityEnum } from './types';
 
-
-export enum TaskPriorityEnum {
-  none = 0,
-  low = 10,
-  medium = 20,
-  high = 90
-}
-
-export interface TaskItem {
-  uuid: string;
-  updated_at: Date;
-  title: string;
-  priority: TaskPriorityEnum;
-  url: string;
-}
 
 
 export abstract class TasksBucketBaseService {
